[feat] commentCount 및 view 카운트 업 설정

+++ src/main/java/developer/domain/community/entity/Community.java
@@ -24,7 +24,7 @@ public class Community extends Auditable {
     @Column(nullable = false)
     private String imgURL;
     @Column
-    private Long view;
+    private int view;
     @Column
     private String category;
     @ElementCollection
